rename auth-config file to auth
