+++ src/routes/v1/auth.route.js
@@ -2,7 +2,7 @@ const express = require('express');
 const validate = require('../../middlewares/validate');
 const authValidation = require('../../validations/auth.validation');
 const authController = require('../../controllers/auth.controller');
-const auth = require('../../middlewares/auth-config');
+const auth = require('../../middlewares/auth');
 
 const router = express.Router();
