+++ src/routes/v1/user.route.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const auth = require('../../middlewares/auth-config');
+const auth = require('../../middlewares/auth');
 const validate = require('../../middlewares/validate');
 const userValidation = require('../../validations/user.validation');
 const userController = require('../../controllers/user.controller');
